+++ src/main/java/com/konggogi/veganlife/comment/service/CommentLikeService.java
@@ -1,19 +1,15 @@
-package com.konggogi.veganlife.like.service;
+package com.konggogi.veganlife.comment.service;
 
 
 import com.konggogi.veganlife.comment.domain.Comment;
-import com.konggogi.veganlife.comment.service.CommentLikeNotifyService;
-import com.konggogi.veganlife.comment.service.CommentQueryService;
 import com.konggogi.veganlife.global.exception.ErrorCode;
 import com.konggogi.veganlife.like.domain.CommentLike;
-import com.konggogi.veganlife.like.domain.PostLike;
 import com.konggogi.veganlife.like.domain.mapper.LikeMapper;
-import com.konggogi.veganlife.like.exception.IllegalLikeStatusException;
 import com.konggogi.veganlife.like.repository.CommentLikeRepository;
-import com.konggogi.veganlife.like.repository.PostLikeRepository;
 import com.konggogi.veganlife.member.domain.Member;
 import com.konggogi.veganlife.member.service.MemberQueryService;
 import com.konggogi.veganlife.post.domain.Post;
+import com.konggogi.veganlife.post.exception.IllegalLikeStatusException;
 import com.konggogi.veganlife.post.service.PostQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,30 +18,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class LikeService {
+public class CommentLikeService {
     private final MemberQueryService memberQueryService;
     private final PostQueryService postQueryService;
     private final CommentQueryService commentQueryService;
-    private final LikeQueryService likeQueryService;
+    private final CommentLikeQueryService commentLikeQueryService;
     private final CommentLikeNotifyService commentLikeNotifyService;
-    private final PostLikeRepository postLikeRepository;
     private final CommentLikeRepository commentLikeRepository;
     private final LikeMapper likeMapper;
-
-    public void addPostLike(Long memberId, Long postId) {
-        Member member = memberQueryService.search(memberId);
-        Post post = postQueryService.search(postId);
-        validatePostLikeIsExist(memberId, postId);
-        PostLike postLike = likeMapper.toPostLike(member);
-        post.addPostLike(postLike);
-    }
-
-    public void removePostLike(Long memberId, Long postId) {
-        memberQueryService.search(memberId);
-        Post post = postQueryService.search(postId);
-        PostLike postLike = validatePostLikeIsNotExist(memberId, postId);
-        post.removePostLike(postLike);
-    }
 
     public void addCommentLike(Long memberId, Long postId, Long commentId) {
         Member member = memberQueryService.search(memberId);
@@ -65,31 +45,12 @@ public class LikeService {
         comment.removeCommentLike(commentLike);
     }
 
-    public void removeMemberFromLike(Long memberId) {
-        postLikeRepository.setMemberToNull(memberId);
+    public void removeMemberFromCommentLike(Long memberId) {
         commentLikeRepository.setMemberToNull(memberId);
     }
 
-    private void validatePostLikeIsExist(Long memberId, Long postId) {
-        likeQueryService
-                .searchPostLike(memberId, postId)
-                .ifPresent(
-                        postLike -> {
-                            throw new IllegalLikeStatusException(ErrorCode.ALREADY_POST_LIKED);
-                        });
-    }
-
-    private PostLike validatePostLikeIsNotExist(Long memberId, Long postId) {
-        return likeQueryService
-                .searchPostLike(memberId, postId)
-                .orElseThrow(
-                        () -> {
-                            throw new IllegalLikeStatusException(ErrorCode.ALREADY_POST_UNLIKED);
-                        });
-    }
-
     private void validateCommentLikeIsExist(Long memberId, Long commentId) {
-        likeQueryService
+        commentLikeQueryService
                 .searchCommentLike(memberId, commentId)
                 .ifPresent(
                         postLike -> {
@@ -98,7 +59,7 @@ public class LikeService {
     }
 
     private CommentLike validateCommentLikeIsNotExist(Long memberId, Long commentId) {
-        return likeQueryService
+        return commentLikeQueryService
                 .searchCommentLike(memberId, commentId)
                 .orElseThrow(
                         () -> {
